Add chapter 5

+++ ch5/strlen.c
@@ -1,5 +1,5 @@
 /* strlen: return leangth of a string by address arithmetic */
-int strlen(char* s) {
+size_t strlen(char* s) {
     char* p = s;
     while (*p != '\0')
         ++p;
